#308 Fixed missing counts caused by a Hibernate 3.2 change server-side.
git-svn-id: file:///home/svn/omero/trunk@899 05709c45-44f0-0310-885b-81a1db45b4a6

+++ components/client/src/pojos/DataObject.java
@@ -296,7 +296,7 @@ public abstract class DataObject
         return asIObject().getDetails();
     }
     
-    protected Integer getCount( String countName )
+    protected Long getCount( String countName )
     {
         Object count = null;
         
@@ -304,8 +304,8 @@ public abstract class DataObject
                 && getDetails().getCounts() != null)
             count = getDetails().getCounts().get( countName );
     
-        if ( count instanceof Integer )
-            return (Integer) count;
+        if ( count instanceof Long )
+            return (Long) count;
 
         return null;
